+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/config/general/GeneralConfiguration.kt
@@ -25,6 +25,8 @@ class GeneralConfiguration {
 
   val folders = LimitConfiguration()
 
+  val bannerWaypoints = BannerWaypointsConfiguration()
+
   val customIconFilter = CustomIconFilterConfiguration()
 
   val openWithItem = OpenWithItemConfiguration()